+++ lib/layout.dart
@@ -7,11 +7,9 @@
 library dart_web_toolkit_layout;
 
 import 'dart:html' as dart_html;
-import 'dart:math' as dart_math;
 
 import 'animation.dart';
 import 'util.dart';
-import 'scheduler.dart';
 
 part 'src/layout/layout.dart';
 part 'src/layout/impl/layout_impl.dart';