+++ lib/src/role/aria_value_attribute.dart
@@ -6,11 +6,11 @@ part of dart_web_toolkit_role;
 /**
  *  ARIA specific type attribute.
  */
-class AriaValueAttribute<T> extends Attribute<T> {
+class AriaValueAttribute<T extends AriaAttributeType> extends Attribute<T> {
 
   AriaValueAttribute(String name, [String defaultValue = null]) : super(name, defaultValue);
 
   String getSingleValue(T value) {
-    return (value as AriaAttributeType).getAriaValue();
+    return value.getAriaValue();
   }
 }