+++ test/dart_web_toolkit_test.dart
@@ -4,14 +4,10 @@
 library dart_web_toolkit_test;
 
 import 'dart:html' as dart_html;
-import 'dart:async' as dart_async;
 import 'package:unittest/unittest.dart';
-import 'package:unittest/mock.dart';
 
 import 'package:dart_web_toolkit/event.dart' as event;
 import 'package:dart_web_toolkit/ui.dart' as ui;
-import 'package:dart_web_toolkit/util.dart' as util;
-import 'package:dart_web_toolkit/i18n.dart' as i18n;
 
 part 'test_group.dart';
 part 'src/solid_ui_object.dart';